move tests for deprecated contrib.metrics
