+++ tests/ignite/metrics/regression/test_fractional_bias.py
@@ -5,12 +5,21 @@ import pytest
 import torch
 
 import ignite.distributed as idist
-from ignite.contrib.metrics.regression import WaveHedgesDistance
 from ignite.engine import Engine
+from ignite.exceptions import NotComputableError
+from ignite.metrics.regression import FractionalBias
+
+
+def test_zero_sample():
+    m = FractionalBias()
+    with pytest.raises(
+        NotComputableError, match=r"FractionalBias must have at least one example before it can be computed"
+    ):
+        m.compute()
 
 
 def test_wrong_input_shapes():
-    m = WaveHedgesDistance()
+    m = FractionalBias()
 
     with pytest.raises(ValueError, match=r"Input data shapes should be the same, but given"):
         m.update((torch.rand(4), torch.rand(4, 1)))
@@ -19,30 +28,38 @@ def test_wrong_input_shapes():
         m.update((torch.rand(4, 1), torch.rand(4)))
 
 
-def test_compute():
+def test_fractional_bias():
     a = np.random.randn(4)
     b = np.random.randn(4)
     c = np.random.randn(4)
     d = np.random.randn(4)
     ground_truth = np.random.randn(4)
 
-    m = WaveHedgesDistance()
+    m = FractionalBias()
 
     m.update((torch.from_numpy(a), torch.from_numpy(ground_truth)))
-    np_sum = (np.abs(ground_truth - a) / np.maximum.reduce([a, ground_truth])).sum()
-    assert m.compute() == pytest.approx(np_sum)
+    np_sum = (2 * (ground_truth - a) / (a + ground_truth)).sum()
+    np_len = len(a)
+    np_ans = np_sum / np_len
+    assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(b), torch.from_numpy(ground_truth)))
-    np_sum += (np.abs(ground_truth - b) / np.maximum.reduce([b, ground_truth])).sum()
-    assert m.compute() == pytest.approx(np_sum)
+    np_sum += (2 * (ground_truth - b) / (b + ground_truth)).sum()
+    np_len += len(b)
+    np_ans = np_sum / np_len
+    assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(c), torch.from_numpy(ground_truth)))
-    np_sum += (np.abs(ground_truth - c) / np.maximum.reduce([c, ground_truth])).sum()
-    assert m.compute() == pytest.approx(np_sum)
+    np_sum += (2 * (ground_truth - c) / (c + ground_truth)).sum()
+    np_len += len(c)
+    np_ans = np_sum / np_len
+    assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(d), torch.from_numpy(ground_truth)))
-    np_sum += (np.abs(ground_truth - d) / np.maximum.reduce([d, ground_truth])).sum()
-    assert m.compute() == pytest.approx(np_sum)
+    np_sum += (2 * (ground_truth - d) / (d + ground_truth)).sum()
+    np_len += len(d)
+    np_ans = np_sum / np_len
+    assert m.compute() == pytest.approx(np_ans)
 
 
 def test_integration():
@@ -55,18 +72,20 @@ def test_integration():
 
         engine = Engine(update_fn)
 
-        m = WaveHedgesDistance()
-        m.attach(engine, "whd")
+        m = FractionalBias()
+        m.attach(engine, "fb")
 
-        np_y = y.numpy().ravel()
-        np_y_pred = y_pred.numpy().ravel()
+        np_y = y.double().numpy().ravel()
+        np_y_pred = y_pred.double().numpy().ravel()
 
         data = list(range(y_pred.shape[0] // batch_size))
-        whd = engine.run(data, max_epochs=1).metrics["whd"]
+        fb = engine.run(data, max_epochs=1).metrics["fb"]
 
-        np_sum = (np.abs(np_y - np_y_pred) / np.maximum.reduce([np_y_pred, np_y])).sum()
+        np_sum = (2 * (np_y - np_y_pred) / (np_y_pred + np_y)).sum()
+        np_len = len(y_pred)
+        np_ans = np_sum / np_len
 
-        assert np_sum == pytest.approx(whd)
+        assert np_ans == pytest.approx(fb)
 
     def get_test_cases():
         test_cases = [
@@ -82,12 +101,18 @@ def test_integration():
             _test(y_pred, y, batch_size)
 
 
-def _test_distrib_compute(device):
+def test_error_is_not_nan():
+    m = FractionalBias()
+    m.update((torch.zeros(4), torch.zeros(4)))
+    assert not (torch.isnan(m._sum_of_errors).any() or torch.isinf(m._sum_of_errors).any()), m._sum_of_errors
+
+
+def _test_distrib_compute(device, tol=1e-5):
     rank = idist.get_rank()
 
     def _test(metric_device):
         metric_device = torch.device(metric_device)
-        m = WaveHedgesDistance(device=metric_device)
+        m = FractionalBias(device=metric_device)
 
         y_pred = torch.randint(0, 10, size=(10,), device=device).float()
         y = torch.randint(0, 10, size=(10,), device=device).float()
@@ -103,9 +128,11 @@ def _test_distrib_compute(device):
 
         res = m.compute()
 
-        np_sum = (np.abs(np_y - np_y_pred) / (np.maximum.reduce([np_y_pred, np_y]) + 1e-30)).sum()
+        np_sum = (2 * (np_y - np_y_pred) / (np_y_pred + np_y + 1e-30)).sum()
+        np_len = len(y_pred)
+        np_ans = np_sum / np_len
 
-        assert np_sum == pytest.approx(res)
+        assert np_ans == pytest.approx(res, rel=tol)
 
     for i in range(3):
         torch.manual_seed(10 + rank + i)
@@ -114,7 +141,7 @@ def _test_distrib_compute(device):
             _test(idist.device())
 
 
-def _test_distrib_integration(device):
+def _test_distrib_integration(device, tol=1e-5):
     rank = idist.get_rank()
 
     def _test(n_epochs, metric_device):
@@ -122,8 +149,8 @@ def _test_distrib_integration(device):
         n_iters = 80
         batch_size = 16
 
-        y_true = torch.rand(size=(n_iters * batch_size,)).to(device)
-        y_preds = torch.rand(size=(n_iters * batch_size,)).to(device)
+        y_true = torch.rand(size=(n_iters * batch_size,), dtype=torch.double).to(device)
+        y_preds = torch.rand(size=(n_iters * batch_size,), dtype=torch.double).to(device)
 
         def update(engine, i):
             return (
@@ -133,8 +160,8 @@ def _test_distrib_integration(device):
 
         engine = Engine(update)
 
-        m = WaveHedgesDistance(device=metric_device)
-        m.attach(engine, "whm")
+        m = FractionalBias(device=metric_device)
+        m.attach(engine, "fb")
 
         data = list(range(n_iters))
         engine.run(data=data, max_epochs=n_epochs)
@@ -142,16 +169,20 @@ def _test_distrib_integration(device):
         y_preds = idist.all_gather(y_preds)
         y_true = idist.all_gather(y_true)
 
-        assert "whm" in engine.state.metrics
+        assert "fb" in engine.state.metrics
 
-        res = engine.state.metrics["whm"]
+        res = engine.state.metrics["fb"]
+        if isinstance(res, torch.Tensor):
+            res = res.cpu().numpy()
 
         np_y_true = y_true.cpu().numpy()
         np_y_preds = y_preds.cpu().numpy()
 
-        np_sum = (np.abs(np_y_true - np_y_preds) / (np.maximum.reduce([np_y_preds, np_y_true]) + 1e-30)).sum()
+        np_sum = (2 * (np_y_true - np_y_preds) / (np_y_preds + np_y_true + 1e-30)).sum()
+        np_len = len(y_preds)
+        np_ans = np_sum / np_len
 
-        assert pytest.approx(res) == np_sum
+        assert pytest.approx(res, rel=tol) == np_ans
 
     metric_devices = ["cpu"]
     if device.type != "xla":
@@ -214,14 +245,14 @@ def test_multinode_distrib_nccl_gpu(distributed_context_multi_node_nccl):
 @pytest.mark.skipif(not idist.has_xla_support, reason="Skip if no PyTorch XLA package")
 def test_distrib_single_device_xla():
     device = idist.device()
-    _test_distrib_compute(device)
-    _test_distrib_integration(device)
+    _test_distrib_compute(device, tol=1e-4)
+    _test_distrib_integration(device, tol=1e-4)
 
 
 def _test_distrib_xla_nprocs(index):
     device = idist.device()
-    _test_distrib_compute(device)
-    _test_distrib_integration(device)
+    _test_distrib_compute(device, tol=1e-4)
+    _test_distrib_integration(device, tol=1e-4)
 
 
 @pytest.mark.tpu
