+++ tests/ignite/metrics/regression/test_geometric_mean_absolute_error.py
@@ -5,21 +5,21 @@ import pytest
 import torch
 
 import ignite.distributed as idist
-from ignite.contrib.metrics.regression import FractionalBias
 from ignite.engine import Engine
 from ignite.exceptions import NotComputableError
+from ignite.metrics.regression import GeometricMeanAbsoluteError
 
 
 def test_zero_sample():
-    m = FractionalBias()
+    m = GeometricMeanAbsoluteError()
     with pytest.raises(
-        NotComputableError, match=r"FractionalBias must have at least one example before it can be computed"
+        NotComputableError, match=r"GeometricMeanAbsoluteError must have at least one example before it can be computed"
     ):
         m.compute()
 
 
 def test_wrong_input_shapes():
-    m = FractionalBias()
+    m = GeometricMeanAbsoluteError()
 
     with pytest.raises(ValueError, match=r"Input data shapes should be the same, but given"):
         m.update((torch.rand(4), torch.rand(4, 1)))
@@ -28,37 +28,42 @@ def test_wrong_input_shapes():
         m.update((torch.rand(4, 1), torch.rand(4)))
 
 
-def test_fractional_bias():
+def test_compute():
     a = np.random.randn(4)
     b = np.random.randn(4)
     c = np.random.randn(4)
     d = np.random.randn(4)
     ground_truth = np.random.randn(4)
+    np_prod = 1.0
 
-    m = FractionalBias()
-
+    m = GeometricMeanAbsoluteError()
     m.update((torch.from_numpy(a), torch.from_numpy(ground_truth)))
-    np_sum = (2 * (ground_truth - a) / (a + ground_truth)).sum()
+
+    errors = np.abs(ground_truth - a)
+    np_prod = np.multiply.reduce(errors) * np_prod
     np_len = len(a)
-    np_ans = np_sum / np_len
+    np_ans = np.power(np_prod, 1.0 / np_len)
     assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(b), torch.from_numpy(ground_truth)))
-    np_sum += (2 * (ground_truth - b) / (b + ground_truth)).sum()
+    errors = np.abs(ground_truth - b)
+    np_prod = np.multiply.reduce(errors) * np_prod
     np_len += len(b)
-    np_ans = np_sum / np_len
+    np_ans = np.power(np_prod, 1.0 / np_len)
     assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(c), torch.from_numpy(ground_truth)))
-    np_sum += (2 * (ground_truth - c) / (c + ground_truth)).sum()
+    errors = np.abs(ground_truth - c)
+    np_prod = np.multiply.reduce(errors) * np_prod
     np_len += len(c)
-    np_ans = np_sum / np_len
+    np_ans = np.power(np_prod, 1.0 / np_len)
     assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(d), torch.from_numpy(ground_truth)))
-    np_sum += (2 * (ground_truth - d) / (d + ground_truth)).sum()
+    errors = np.abs(ground_truth - d)
+    np_prod = np.multiply.reduce(errors) * np_prod
     np_len += len(d)
-    np_ans = np_sum / np_len
+    np_ans = np.power(np_prod, 1.0 / np_len)
     assert m.compute() == pytest.approx(np_ans)
 
 
@@ -72,20 +77,20 @@ def test_integration():
 
         engine = Engine(update_fn)
 
-        m = FractionalBias()
-        m.attach(engine, "fb")
+        m = GeometricMeanAbsoluteError()
+        m.attach(engine, "gmae")
 
-        np_y = y.double().numpy().ravel()
-        np_y_pred = y_pred.double().numpy().ravel()
+        np_y = y.numpy().ravel()
+        np_y_pred = y_pred.numpy().ravel()
 
         data = list(range(y_pred.shape[0] // batch_size))
-        fb = engine.run(data, max_epochs=1).metrics["fb"]
+        gmae = engine.run(data, max_epochs=1).metrics["gmae"]
 
-        np_sum = (2 * (np_y - np_y_pred) / (np_y_pred + np_y)).sum()
+        sum_errors = (np.log(np.abs(np_y - np_y_pred))).sum()
         np_len = len(y_pred)
-        np_ans = np_sum / np_len
+        np_ans = np.exp(sum_errors / np_len)
 
-        assert np_ans == pytest.approx(fb)
+        assert np_ans == pytest.approx(gmae)
 
     def get_test_cases():
         test_cases = [
@@ -94,25 +99,21 @@ def test_integration():
         ]
         return test_cases
 
-    for _ in range(5):
+    for i in range(5):
         # check multiple random inputs as random exact occurencies are rare
+        torch.manual_seed(12 + i)
         test_cases = get_test_cases()
         for y_pred, y, batch_size in test_cases:
             _test(y_pred, y, batch_size)
 
 
-def test_error_is_not_nan():
-    m = FractionalBias()
-    m.update((torch.zeros(4), torch.zeros(4)))
-    assert not (torch.isnan(m._sum_of_errors).any() or torch.isinf(m._sum_of_errors).any()), m._sum_of_errors
-
-
-def _test_distrib_compute(device, tol=1e-5):
+def _test_distrib_compute(device):
     rank = idist.get_rank()
 
     def _test(metric_device):
         metric_device = torch.device(metric_device)
-        m = FractionalBias(device=metric_device)
+        m = GeometricMeanAbsoluteError(device=metric_device)
+        torch.manual_seed(10 + rank)
 
         y_pred = torch.randint(0, 10, size=(10,), device=device).float()
         y = torch.randint(0, 10, size=(10,), device=device).float()
@@ -128,20 +129,19 @@ def _test_distrib_compute(device, tol=1e-5):
 
         res = m.compute()
 
-        np_sum = (2 * (np_y - np_y_pred) / (np_y_pred + np_y + 1e-30)).sum()
+        sum_errors = (np.log(np.abs(np_y - np_y_pred))).sum()
         np_len = len(y_pred)
-        np_ans = np_sum / np_len
+        np_ans = np.exp(sum_errors / np_len)
 
-        assert np_ans == pytest.approx(res, rel=tol)
+        assert np_ans == pytest.approx(res)
 
-    for i in range(3):
-        torch.manual_seed(10 + rank + i)
+    for _ in range(3):
         _test("cpu")
         if device.type != "xla":
             _test(idist.device())
 
 
-def _test_distrib_integration(device, tol=1e-5):
+def _test_distrib_integration(device):
     rank = idist.get_rank()
 
     def _test(n_epochs, metric_device):
@@ -149,8 +149,8 @@ def _test_distrib_integration(device, tol=1e-5):
         n_iters = 80
         batch_size = 16
 
-        y_true = torch.rand(size=(n_iters * batch_size,), dtype=torch.double).to(device)
-        y_preds = torch.rand(size=(n_iters * batch_size,), dtype=torch.double).to(device)
+        y_true = torch.rand(size=(n_iters * batch_size,)).to(device)
+        y_preds = torch.rand(size=(n_iters * batch_size,)).to(device)
 
         def update(engine, i):
             return (
@@ -160,8 +160,8 @@ def _test_distrib_integration(device, tol=1e-5):
 
         engine = Engine(update)
 
-        m = FractionalBias(device=metric_device)
-        m.attach(engine, "fb")
+        m = GeometricMeanAbsoluteError(device=metric_device)
+        m.attach(engine, "gmae")
 
         data = list(range(n_iters))
         engine.run(data=data, max_epochs=n_epochs)
@@ -169,27 +169,25 @@ def _test_distrib_integration(device, tol=1e-5):
         y_preds = idist.all_gather(y_preds)
         y_true = idist.all_gather(y_true)
 
-        assert "fb" in engine.state.metrics
+        assert "gmae" in engine.state.metrics
 
-        res = engine.state.metrics["fb"]
-        if isinstance(res, torch.Tensor):
-            res = res.cpu().numpy()
+        res = engine.state.metrics["gmae"]
 
         np_y_true = y_true.cpu().numpy()
         np_y_preds = y_preds.cpu().numpy()
 
-        np_sum = (2 * (np_y_true - np_y_preds) / (np_y_preds + np_y_true + 1e-30)).sum()
+        sum_errors = (np.log(np.abs(np_y_true - np_y_preds))).sum()
         np_len = len(y_preds)
-        np_ans = np_sum / np_len
+        np_ans = np.exp(sum_errors / np_len)
 
-        assert pytest.approx(res, rel=tol) == np_ans
+        assert pytest.approx(res) == np_ans
 
     metric_devices = ["cpu"]
     if device.type != "xla":
         metric_devices.append(idist.device())
     for metric_device in metric_devices:
         for i in range(2):
-            torch.manual_seed(12 + rank + i)
+            torch.manual_seed(11 + rank + i)
             _test(n_epochs=1, metric_device=metric_device)
             _test(n_epochs=2, metric_device=metric_device)
 
@@ -226,7 +224,7 @@ def test_distrib_hvd(gloo_hvd_executor):
 @pytest.mark.skipif(not idist.has_native_dist_support, reason="Skip if no native dist support")
 @pytest.mark.skipif("MULTINODE_DISTRIB" not in os.environ, reason="Skip if not multi-node distributed")
 def test_multinode_distrib_gloo_cpu_or_gpu(distributed_context_multi_node_gloo):
-    device = idist.device()
+    device = torch.device("cpu" if not torch.cuda.is_available() else "cuda")
     _test_distrib_compute(device)
     _test_distrib_integration(device)
 
@@ -245,14 +243,14 @@ def test_multinode_distrib_nccl_gpu(distributed_context_multi_node_nccl):
 @pytest.mark.skipif(not idist.has_xla_support, reason="Skip if no PyTorch XLA package")
 def test_distrib_single_device_xla():
     device = idist.device()
-    _test_distrib_compute(device, tol=1e-4)
-    _test_distrib_integration(device, tol=1e-4)
+    _test_distrib_compute(device)
+    _test_distrib_integration(device)
 
 
 def _test_distrib_xla_nprocs(index):
     device = idist.device()
-    _test_distrib_compute(device, tol=1e-4)
-    _test_distrib_integration(device, tol=1e-4)
+    _test_distrib_compute(device)
+    _test_distrib_integration(device)
 
 
 @pytest.mark.tpu
