+++ tests/ignite/metrics/regression/test_manhattan_distance.py
@@ -3,133 +3,127 @@ import os
 import numpy as np
 import pytest
 import torch
+from sklearn.metrics import DistanceMetric
 
 import ignite.distributed as idist
-from ignite.contrib.metrics.regression import MedianAbsoluteError
 from ignite.engine import Engine
-from ignite.exceptions import NotComputableError
-
-
-def test_zero_sample():
-    m = MedianAbsoluteError()
-    with pytest.raises(
-        NotComputableError, match=r"EpochMetric must have at least one example before it can be computed"
-    ):
-        m.compute()
+from ignite.metrics.regression import ManhattanDistance
 
 
 def test_wrong_input_shapes():
-    m = MedianAbsoluteError()
+    m = ManhattanDistance()
 
-    with pytest.raises(ValueError, match=r"Predictions should be of shape"):
-        m.update((torch.rand(4, 1, 2), torch.rand(4, 1)))
+    with pytest.raises(ValueError, match=r"Input data shapes should be the same, but given"):
+        m.update((torch.rand(4), torch.rand(4, 1)))
 
-    with pytest.raises(ValueError, match=r"Targets should be of shape"):
-        m.update((torch.rand(4, 1), torch.rand(4, 1, 2)))
-
-    with pytest.raises(ValueError, match=r"Predictions should be of shape"):
-        m.update((torch.rand(4, 1, 2), torch.rand(4)))
-
-    with pytest.raises(ValueError, match=r"Targets should be of shape"):
-        m.update((torch.rand(4), torch.rand(4, 1, 2)))
+    with pytest.raises(ValueError, match=r"Input data shapes should be the same, but given"):
+        m.update((torch.rand(4, 1), torch.rand(4)))
 
 
-def test_median_absolute_error():
-    # See https://github.com/torch/torch7/pull/182
-    # For even number of elements, PyTorch returns middle element
-    # NumPy returns average of middle elements
-    # Size of dataset will be odd for these tests
+def test_mahattan_distance():
+    a = np.random.randn(4)
+    b = np.random.randn(4)
+    c = np.random.randn(4)
+    d = np.random.randn(4)
+    ground_truth = np.random.randn(4)
 
-    size = 51
-    np_y_pred = np.random.rand(size)
-    np_y = np.random.rand(size)
-    np_median_absolute_error = np.median(np.abs(np_y - np_y_pred))
+    m = ManhattanDistance()
 
-    m = MedianAbsoluteError()
-    y_pred = torch.from_numpy(np_y_pred)
-    y = torch.from_numpy(np_y)
+    manhattan = DistanceMetric.get_metric("manhattan")
 
-    m.reset()
-    m.update((y_pred, y))
+    m.update((torch.from_numpy(a), torch.from_numpy(ground_truth)))
+    np_sum = np.abs(ground_truth - a).sum()
+    assert m.compute() == pytest.approx(np_sum)
+    assert manhattan.pairwise([a, ground_truth])[0][1] == pytest.approx(np_sum)
 
-    assert np_median_absolute_error == pytest.approx(m.compute())
+    m.update((torch.from_numpy(b), torch.from_numpy(ground_truth)))
+    np_sum += np.abs(ground_truth - b).sum()
+    assert m.compute() == pytest.approx(np_sum)
+    v1 = np.hstack([a, b])
+    v2 = np.hstack([ground_truth, ground_truth])
+    assert manhattan.pairwise([v1, v2])[0][1] == pytest.approx(np_sum)
 
+    m.update((torch.from_numpy(c), torch.from_numpy(ground_truth)))
+    np_sum += np.abs(ground_truth - c).sum()
+    assert m.compute() == pytest.approx(np_sum)
+    v1 = np.hstack([v1, c])
+    v2 = np.hstack([v2, ground_truth])
+    assert manhattan.pairwise([v1, v2])[0][1] == pytest.approx(np_sum)
 
-def test_median_absolute_error_2():
-    np.random.seed(1)
-    size = 105
-    np_y_pred = np.random.rand(size, 1)
-    np_y = np.random.rand(size, 1)
-    np.random.shuffle(np_y)
-    np_median_absolute_error = np.median(np.abs(np_y - np_y_pred))
-
-    m = MedianAbsoluteError()
-    y_pred = torch.from_numpy(np_y_pred)
-    y = torch.from_numpy(np_y)
-
-    m.reset()
-    batch_size = 16
-    n_iters = size // batch_size + 1
-    for i in range(n_iters):
-        idx = i * batch_size
-        m.update((y_pred[idx : idx + batch_size], y[idx : idx + batch_size]))
-
-    assert np_median_absolute_error == pytest.approx(m.compute())
+    m.update((torch.from_numpy(d), torch.from_numpy(ground_truth)))
+    np_sum += np.abs(ground_truth - d).sum()
+    assert m.compute() == pytest.approx(np_sum)
+    v1 = np.hstack([v1, d])
+    v2 = np.hstack([v2, ground_truth])
+    assert manhattan.pairwise([v1, v2])[0][1] == pytest.approx(np_sum)
 
 
-def test_integration_median_absolute_error():
-    np.random.seed(1)
-    size = 105
-    np_y_pred = np.random.rand(size, 1)
-    np_y = np.random.rand(size, 1)
-    np.random.shuffle(np_y)
-    np_median_absolute_error = np.median(np.abs(np_y - np_y_pred))
+def test_integration():
+    def _test(y_pred, y, batch_size):
+        def update_fn(engine, batch):
+            idx = (engine.state.iteration - 1) * batch_size
+            y_true_batch = np_y[idx : idx + batch_size]
+            y_pred_batch = np_y_pred[idx : idx + batch_size]
+            return torch.from_numpy(y_pred_batch), torch.from_numpy(y_true_batch)
 
-    batch_size = 15
+        engine = Engine(update_fn)
 
-    def update_fn(engine, batch):
-        idx = (engine.state.iteration - 1) * batch_size
-        y_true_batch = np_y[idx : idx + batch_size]
-        y_pred_batch = np_y_pred[idx : idx + batch_size]
-        return torch.from_numpy(y_pred_batch), torch.from_numpy(y_true_batch)
+        m = ManhattanDistance()
+        m.attach(engine, "md")
 
-    engine = Engine(update_fn)
+        np_y = y.numpy().ravel()
+        np_y_pred = y_pred.numpy().ravel()
 
-    m = MedianAbsoluteError()
-    m.attach(engine, "median_absolute_error")
+        manhattan = DistanceMetric.get_metric("manhattan")
 
-    data = list(range(size // batch_size))
-    median_absolute_error = engine.run(data, max_epochs=1).metrics["median_absolute_error"]
+        data = list(range(y_pred.shape[0] // batch_size))
+        md = engine.run(data, max_epochs=1).metrics["md"]
 
-    assert np_median_absolute_error == pytest.approx(median_absolute_error)
+        assert manhattan.pairwise([np_y_pred, np_y])[0][1] == pytest.approx(md)
+
+    def get_test_cases():
+        test_cases = [
+            (torch.rand(size=(100,)), torch.rand(size=(100,)), 10),
+            (torch.rand(size=(100, 1)), torch.rand(size=(100, 1)), 20),
+        ]
+        return test_cases
+
+    for _ in range(5):
+        # check multiple random inputs as random exact occurencies are rare
+        test_cases = get_test_cases()
+        for y_pred, y, batch_size in test_cases:
+            _test(y_pred, y, batch_size)
+
+
+def test_error_is_not_nan():
+    m = ManhattanDistance()
+    m.update((torch.zeros(4), torch.zeros(4)))
+    assert not (torch.isnan(m._sum_of_errors).any() or torch.isinf(m._sum_of_errors).any()), m._sum_of_errors
 
 
 def _test_distrib_compute(device):
+    rank = idist.get_rank()
+
+    manhattan = DistanceMetric.get_metric("manhattan")
+
     def _test(metric_device):
         metric_device = torch.device(metric_device)
-        m = MedianAbsoluteError(device=metric_device)
+        m = ManhattanDistance(device=metric_device)
 
-        size = 105
+        y_pred = torch.randint(0, 10, size=(10,), device=device).float()
+        y = torch.randint(0, 10, size=(10,), device=device).float()
 
-        y_pred = torch.randint(1, 10, size=(size, 1), dtype=torch.double, device=device)
-        y = torch.randint(1, 10, size=(size, 1), dtype=torch.double, device=device)
         m.update((y_pred, y))
 
         # gather y_pred, y
         y_pred = idist.all_gather(y_pred)
         y = idist.all_gather(y)
 
-        np_y_pred = y_pred.cpu().numpy().ravel()
-        np_y = y.cpu().numpy().ravel()
-
+        np_y_pred = y_pred.cpu().numpy()
+        np_y = y.cpu().numpy()
         res = m.compute()
+        assert manhattan.pairwise([np_y_pred, np_y])[0][1] == pytest.approx(res)
 
-        e = np.abs(np_y - np_y_pred)
-
-        np_res = np.median(e)
-        assert pytest.approx(res) == np_res
-
-    rank = idist.get_rank()
     for i in range(3):
         torch.manual_seed(10 + rank + i)
         _test("cpu")
@@ -138,10 +132,15 @@ def _test_distrib_compute(device):
 
 
 def _test_distrib_integration(device):
+    rank = idist.get_rank()
+
+    manhattan = DistanceMetric.get_metric("manhattan")
+
     def _test(n_epochs, metric_device):
         metric_device = torch.device(metric_device)
         n_iters = 80
-        batch_size = 105
+        batch_size = 16
+
         y_true = torch.rand(size=(n_iters * batch_size,)).to(device)
         y_preds = torch.rand(size=(n_iters * batch_size,)).to(device)
 
@@ -153,8 +152,8 @@ def _test_distrib_integration(device):
 
         engine = Engine(update)
 
-        m = MedianAbsoluteError(device=metric_device)
-        m.attach(engine, "mae")
+        m = ManhattanDistance(device=metric_device)
+        m.attach(engine, "md")
 
         data = list(range(n_iters))
         engine.run(data=data, max_epochs=n_epochs)
@@ -162,25 +161,23 @@ def _test_distrib_integration(device):
         y_preds = idist.all_gather(y_preds)
         y_true = idist.all_gather(y_true)
 
-        assert "mae" in engine.state.metrics
+        assert "md" in engine.state.metrics
 
-        res = engine.state.metrics["mae"]
+        res = engine.state.metrics["md"]
+        if isinstance(res, torch.Tensor):
+            res = res.cpu().numpy()
 
-        np_y_true = y_true.cpu().numpy().ravel()
-        np_y_preds = y_preds.cpu().numpy().ravel()
+        np_y_true = y_true.cpu().numpy()
+        np_y_preds = y_preds.cpu().numpy()
 
-        e = np.abs(np_y_true - np_y_preds)
-        np_res = np.median(e)
-
-        assert pytest.approx(res) == np_res
+        assert pytest.approx(res) == manhattan.pairwise([np_y_preds, np_y_true])[0][1]
 
     metric_devices = ["cpu"]
     if device.type != "xla":
         metric_devices.append(idist.device())
     for metric_device in metric_devices:
-        rank = idist.get_rank()
         for i in range(2):
-            torch.manual_seed(10 + rank + i)
+            torch.manual_seed(12 + rank + i)
             _test(n_epochs=1, metric_device=metric_device)
             _test(n_epochs=2, metric_device=metric_device)
