+++ tests/ignite/metrics/regression/test_median_relative_absolute_error.py
@@ -5,9 +5,9 @@ import pytest
 import torch
 
 import ignite.distributed as idist
-from ignite.contrib.metrics.regression import MedianRelativeAbsoluteError
 from ignite.engine import Engine
 from ignite.exceptions import NotComputableError
+from ignite.metrics.regression import MedianRelativeAbsoluteError
 
 
 def test_zero_sample():
